requirements.txt
asgiref==3.2.3
bson==0.5.8
dataclasses==0.6
dj-database-url==0.5.0
Django==2.2.9
djangorestframework==3.11.0
djongo==1.3.0
gunicorn==20.0.4
pymongo==3.10.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
six==1.14.0
sqlparse==0.2.4
